+++ fa18Team22/fa18Team22/Controllers/ProcurementSearchController.cs
@@ -49,7 +49,7 @@ namespace fa18Team22.Controllers
                 apvm.ProcurementDate = System.DateTime.Today;
                 apvm.BookID = book.BookID;
                 apvm.Author = book.Author;
-                apvm.AvgRatingProc = book.AvgRating;
+                apvm.AvgRatingProc = (decimal)book.AvgRating;
                 apvm.PublishDate = book.PublishDate;
                 apvm.Cost = book.BookCost;
                 apvm.userID = User.Identity.Name;
@@ -220,7 +220,7 @@ namespace fa18Team22.Controllers
                 apvm.ProcurementDate = System.DateTime.Today;
                 apvm.BookID = book.BookID;
                 apvm.Author = book.Author;
-                apvm.AvgRatingProc = book.AvgRating;
+                apvm.AvgRatingProc = (decimal)book.AvgRating;
                 apvm.PublishDate = book.PublishDate;
                 apvm.Cost = book.BookCost;
                 apvm.userID = User.Identity.Name;
